restrucuturing files in client folder
